Moved some constants for the transformation script to their own files to be imported
