+++ transformer/node_types.py
@@ -1,0 +1,7 @@
+# Enum of node types in the datamodel
+class NODE_TYPES(Enum):
+    DIAGNOSIS = "Diagnosis"
+    PARTICIPANT = "Participant"
+    REFERENCE_FILE = "Reference File"
+    STUDY = "Study"
+    SURVIVAL = "Survival"
